Update Dockerfiles, fix sqrt usage, clean editor submodule
- Add libzstd-dev and set MMO_OPENSSL_STATIC=OFF in all server Dockerfiles.
- Replace std::sqrtf with std::sqrt in creature_ai_combat_state.cpp for distance calculations.
- Update editor submodule reference to a clean state.

+++ src/shared/game_server/ai/creature_ai_combat_state.cpp
@@ -1214,8 +1214,8 @@ namespace mmo
 				const bool inRange = distanceSq <= meleeRangeSq;
 				if (inRange)
 				{
-					const float distance = std::sqrtf(distanceSq);
-					const float acceptanceRadius = std::sqrtf(meleeRangeSq);
+					const float distance = std::sqrt(distanceSq);
+					const float acceptanceRadius = std::sqrt(meleeRangeSq);
 					DLOG("Creature " << controlled.GetGuid() << " stopped at engagement range: distance=" 
 						<< distance << " <= acceptanceRadius=" << acceptanceRadius);
 				}
@@ -1229,7 +1229,7 @@ namespace mmo
 				const bool inRange = distanceSq >= minRangeSq && distanceSq <= maxRangeSq;
 				if (inRange)
 				{
-					const float distance = std::sqrtf(distanceSq);
+					const float distance = std::sqrt(distanceSq);
 					const float acceptanceRadius = CASTER_OPTIMAL_RANGE;
 					DLOG("Creature " << controlled.GetGuid() << " stopped at engagement range: distance=" 
 						<< distance << " <= acceptanceRadius=" << acceptanceRadius);
